requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.0

# Database (PostgreSQL)
sqlalchemy==2.0.35
alembic==1.13.2
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Async Tasks
celery==5.4.0
redis==5.1.0
flower==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
authlib==1.3.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# SMS (Aliyun)
alibabacloud-dysmsapi20170525==2.0.24

# Storage
boto3==1.35.20
oss2==2.18.6

# Payment
stripe==10.12.0
wechatpayv3==1.2.39

# Region Detection
geoip2==4.8.1
maxminddb==2.6.2

# Monitoring & Logging
prometheus-client==0.20.0
structlog==24.4.0
python-json-logger==2.0.7

# Security & Validation
slowapi==0.1.9
email-validator==2.2.0
python-decouple==3.8

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-httpx==0.30.0
factory-boy==3.3.1
faker==28.4.1

# Development Tools
black==24.8.0
isort==5.13.2
mypy==1.11.2
pylint==3.3.0
pre-commit==3.8.0